requirements.txt
anthropic==0.49.0
Flask==3.1.0
gql==3.5.1
matplotlib==3.10.1
numpy==2.2.4
pandas==2.2.3
pgeocode==0.5.0
python-dotenv==1.0.1
seaborn==0.13.2
SQLAlchemy==2.0.37
tqdm==4.67.1
psycopg2-binary==2.9.10
aiohttp==3.11.14
